feat(avalara-integration): change namespace in mutation for fetching provider taxes (#3504)
## Context

Currently Lago is working on integration with Avalara tax provider

## Description

In the one-off invoice form, there is a button for fetching taxes for
the form content. It fetches taxes on the fly and adjust the form.

Logic was namespaced under Anrok, another tax tool.

This PR changes namespace to more generic one, since the mutation itself
should not care about underlying integration
