+++ app/graphql/types/integrations/tax_objects/fee_object.rb
@@ -2,16 +2,16 @@
 
 module Types
   module Integrations
-    module AnrokObjects
+    module TaxObjects
       class FeeObject < Types::BaseObject
-        graphql_name "AnrokFeeObject"
+        graphql_name "TaxFeeObject"
 
         field :amount_cents, GraphQL::Types::BigInt, null: true
         field :item_code, String, null: true
         field :item_id, String, null: true
         field :tax_amount_cents, GraphQL::Types::BigInt, null: true
 
-        field :tax_breakdown, [Types::Integrations::AnrokObjects::BreakdownObject]
+        field :tax_breakdown, [Types::Integrations::TaxObjects::BreakdownObject]
       end
     end
   end
