+++ app/graphql/types/mutation_type.rb
@@ -154,7 +154,7 @@ module Types
     field :create_avalara_integration, mutation: Mutations::Integrations::Avalara::Create
     field :update_avalara_integration, mutation: Mutations::Integrations::Avalara::Update
 
-    field :fetch_draft_invoice_taxes, mutation: Mutations::Integrations::Anrok::FetchDraftInvoiceTaxes
+    field :fetch_draft_invoice_taxes, mutation: Mutations::Integrations::FetchDraftInvoiceTaxes
 
     field :create_xero_integration, mutation: Mutations::Integrations::Xero::Create
     field :update_xero_integration, mutation: Mutations::Integrations::Xero::Update
